% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyGamma.R
\name{applyGamma}
\alias{applyGamma}
\title{Apply Gamma distribution to sample and compute required components for the test.}
\usage{
applyGamma(x, use.rate)
}
\arguments{
\item{x}{a numeric vector.}

\item{use.rate}{logical. If \code{TRUE} the rate parameter is returned while estimating MLE. Otherwise the scale is returned.}
}
\value{
a list with three elements.
}
\description{
Compute Maximum likelihood estimates of the parameters in Gamma distribution, Score function evaluated at the sample,
and probability inverse transformed (PIT) values of sample.
}
