% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyLMNormal.R
\name{applyLMNormal}
\alias{applyLMNormal}
\title{Apply linear model and extract required components for the test}
\usage{
applyLMNormal(x, y)
}
\arguments{
\item{x}{is either a numeric vector or a design matrix. In the design matrix, rows indicate observations and columns
presents explanatory variables.}

\item{y}{is a vector of numeric values with the same number of observations or number of rows as x.}
}
\value{
a list with three elements.
}
\description{
Apply linear model and extract required components for the test
}
