% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_exp.R
\name{expFx}
\alias{expFx}
\title{The Probability Transform Inverse Function for Exponential Distribution}
\usage{
expFx(x, theta)
}
\arguments{
\item{x}{a vector data}

\item{theta}{MLE estimate of data}
}
\value{
A vector of transformed values
}
\description{
The Probability Transform Inverse Function for Exponential Distribution
}
