% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_exp.R
\name{expMLE}
\alias{expMLE}
\title{Calculate MLE of Parameter in Exponential Distribution}
\usage{
expMLE(x)
}
\arguments{
\item{x}{a vector data}
}
\value{
Numeric
}
\description{
Calculate MLE of Parameter in Exponential Distribution
}
