% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_exp.R
\name{expPIT}
\alias{expPIT}
\title{Compute probability inverse transform values for Exponential dist.}
\usage{
expPIT(x, theta)
}
\arguments{
\item{x}{a numeric vector of length n}

\item{theta}{a numeric vector of length one, containing MLE rate in Exponential dist.}
}
\value{
a vector of length n containing the probability inverse transformed (PIT) values
}
\description{
Compute probability inverse transform values for Exponential dist.
}
