% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_exp.R
\name{expScore}
\alias{expScore}
\title{Compute score function for Exponential distribution.}
\usage{
expScore(x, theta)
}
\arguments{
\item{x}{a numeric vector of length n}

\item{theta}{a numeric vector of length two, containing MLE of parameters in Exponential dist.}
}
\value{
Score matrix with n rows and two columns.
}
\description{
Compute score function for Exponential distribution.
}
