% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gamma.R
\name{gammaFisherByHessian}
\alias{gammaFisherByHessian}
\title{Compute Fisher information matrix by the negative expected value of Hessian matrix in Gamma distribution.}
\usage{
gammaFisherByHessian(theta)
}
\arguments{
\item{theta}{a numeric vector of length two, containing MLE of parameters in Gamma dist}
}
\value{
Fisher information matrix for Gamma distribution
}
\description{
Compute Fisher information matrix by the negative expected value of Hessian matrix in Gamma distribution.
}
