% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gamma.R
\name{gammaMLE}
\alias{gammaMLE}
\title{Compute maximum likelihood estimate of shape and scale parameter in Gamma distribution.}
\usage{
gammaMLE(x, ur)
}
\arguments{
\item{x}{a numeric vector of length n}

\item{ur}{logical. If \code{TRUE} the rate parameter is returned. Otherwise the scale is returned.}
}
\value{
a vector of length two with shape and scale/rate.
}
\description{
Estimate the MLE of shape and scale parameters of the Gamma distribution using the
Newton-Raphson method on the profile log-likelihood to estimate the shape parameter.
}
