% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gamma.R
\name{gammaPIT}
\alias{gammaPIT}
\title{Compute probability inverse transform values for Gamma distribution}
\usage{
gammaPIT(x, theta)
}
\arguments{
\item{x}{a numeric vector of length n}

\item{theta}{a numeric vector of length two, containing MLE of parameters in Gamma dist.}
}
\value{
a vector of length n containing the probability inverse transformed (PIT) values
}
\description{
Compute probability inverse transform values for Gamma distribution
}
