% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gamma.R
\name{gammaScore}
\alias{gammaScore}
\title{Compute score function for Gamma distribution.}
\usage{
gammaScore(x, theta)
}
\arguments{
\item{x}{a numeric vector of length n}

\item{theta}{a numeric vector of length two, containing MLE of parameters in Gamma dist.}
}
\value{
Score matrix with n rows and two columns.
}
\description{
Compute score function for Gamma distribution.
}
