% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_glmgamma.R
\name{glmMLE}
\alias{glmMLE}
\title{Compute maximum likelihood estimates for a generalized linear model with Gamma response.}
\usage{
glmMLE(fit)
}
\arguments{
\item{fit}{is an object of class \code{glm} and its default value is NULL. If a fit of class \code{glm} is provided,
the arguments \code{x}, \code{y}, and \code{l} will be ignored. We recommend using \code{\link{glm2}} function from
\code{\link{glm2}} package since it provides better convergence while optimizing the likelihood to estimate
coefficients of the model by IWLS method. It is required to return design matrix by \code{x} = \code{TRUE} in
\code{\link{glm}} or \code{\link{glm2}} function. For more information on how to do this, refer to the help
documentation for the \code{\link{glm}} or \code{\link{glm2}} function.}
}
\value{
a numeric vector of estimates.
}
\description{
Compute maximum likelihood estimates for a generalized linear model with Gamma response.
}
