% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_glmgamma.R
\name{glmScorePIT}
\alias{glmScorePIT}
\title{Compute score function for a generalized linear model with Gamma response.}
\usage{
glmScorePIT(fit, theta)
}
\arguments{
\item{fit}{TBD}

\item{theta}{a numeric vector of length (p+1), containing MLE of parameters in a linear model.}
}
\value{
Score matrix with n rows and (p+2) columns.
}
\description{
Compute score function for a generalized linear model with Gamma response.
}
