% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_testYourModel.R
\name{IG_pitfunc}
\alias{IG_pitfunc}
\title{Compute the probability transformed values for a sample from Inverse Gaussian distribution.}
\usage{
IG_pitfunc(obs, ...)
}
\arguments{
\item{obs}{A numeric vector of sample observations.}

\item{...}{A list of additional parameters to define the likelihood. In this function, MLE is being passed.}
}
\value{
A numeric vector of probability transformed values of sample observations.
}
\description{
This function is used in \code{\link{testYourModel}} function for example purposes.
}
