% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_testYourModel.R
\name{IG_scorefunc}
\alias{IG_scorefunc}
\title{Compute the score function of the Inverse Gaussian distribution based on a sample.}
\usage{
IG_scorefunc(obs, ...)
}
\arguments{
\item{obs}{a numeric vector of sample observations.}

\item{...}{a list of additional parameters to define the likelihood. In this function, weight and mle are being passed.}
}
\value{
The score matrix with n rows (number of sample observations) and 2 columns (mean and shape).
}
\description{
Compute the score function of the Inverse Gaussian distribution based on a sample.
}
