% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_testYourModel.R
\name{IGPIT}
\alias{IGPIT}
\title{Compute the probability integral transformed values for a sample from Inverse Gaussian distribution.}
\usage{
IGPIT(obs, ...)
}
\arguments{
\item{obs}{A numeric vector of sample observations.}

\item{...}{A list of additional parameters to define the likelihood.}
}
\value{
A numeric vector of probability integral transformed values of sample observations.
}
\description{
This function is used in \code{\link{testYourModel}} function for example purposes.
}
