% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_testYourModel.R
\name{inversegaussianMLE}
\alias{inversegaussianMLE}
\title{Compute the maximum likelihood estimate of parameters in Inverse Gaussian distribution with weighted observations.}
\usage{
inversegaussianMLE(obs, ...)
}
\arguments{
\item{obs}{a numeric vector of sample observations.}

\item{...}{a list of additional parameters to define the likelihood.}
}
\value{
The function compute the MLE of parameters in Inverse Gaussian distribution and returns a vector of
estimates. The first and second elements of the vector are MLE of the mean and shape, respectively.
}
\description{
This function is used in \code{\link{testYourModel}} function for example purposes.
}
