% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_testYourModel.R
\name{inversegaussianScore}
\alias{inversegaussianScore}
\title{Compute the score function of the Inverse Gaussian distribution based on a sample.}
\usage{
inversegaussianScore(obs, ...)
}
\arguments{
\item{obs}{a numeric vector of sample observations.}

\item{...}{a list of additional parameters to define the likelihood.}
}
\value{
The score matrix with n rows (number of sample observations) and 2 columns (mean and shape).
}
\description{
This function is used in \code{\link{testYourModel}} function for example purposes.
}
