% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lmnormal.R
\name{lmFisherByHessian}
\alias{lmFisherByHessian}
\title{Compute Fisher information matrix in the case of linear model with Normal residuals.}
\usage{
lmFisherByHessian(x, y, theta)
}
\arguments{
\item{x}{a matrix with n rows and p columns containing the explanatory variables.}

\item{y}{a numeric vector of length n containing the response variable.}

\item{theta}{a numeric vector of length (p+1), containing MLE of parameters in a linear model.}
}
\value{
Fisher information matrix for linear models.
}
\description{
Compute Fisher information matrix in the case of linear model with Normal residuals.
}
