% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lmnormal.R
\name{lmMLE}
\alias{lmMLE}
\title{Compute maximum likelihood estimates for linear models}
\usage{
lmMLE(x, y)
}
\arguments{
\item{x}{a matrix with n rows and p columns containing the explanatory variables.}

\item{y}{a numeric vector of length n containing the response variable.}
}
\value{
a numeric vector of estimates.
}
\description{
Compute maximum likelihood estimates for linear models
}
