% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lmnormal.R
\name{lmPIT}
\alias{lmPIT}
\title{Compute probability inverse transform values for linear models.}
\usage{
lmPIT(x, y, theta)
}
\arguments{
\item{x}{a matrix with n rows and p columns containing the explanatory variables.}

\item{y}{a numeric vector of length n containing the response variable.}

\item{theta}{a numeric vector of length (p+1), containing MLE of parameters in a linear model.}
}
\value{
a vector of length n containing the probability inverse transformed (PIT) values
}
\description{
Compute probability inverse transform values for linear models.
}
