% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lmnormal.R
\name{lmScore}
\alias{lmScore}
\title{Compute score function for linear models.}
\usage{
lmScore(x, y, theta)
}
\arguments{
\item{x}{a matrix with n rows and p columns containing the explanatory variables.}

\item{y}{a numeric vector of length n containing the response variable.}

\item{theta}{a numeric vector of length (p+1), containing MLE of parameters in a linear model.}
}
\value{
Score matrix with n rows and (p+1) columns.
}
\description{
Compute score function for linear models.
}
