% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normal.R
\name{normalFisherByHessian}
\alias{normalFisherByHessian}
\title{Compute Fisher information matrix by the negative expected value of Hessian matrix in Normal distribution.}
\usage{
normalFisherByHessian(theta)
}
\arguments{
\item{theta}{a numeric vector of length two, containing MLE of parameters in Normal dist}
}
\value{
Fisher information matrix for Normal distribution
}
\description{
Compute Fisher information matrix by the negative expected value of Hessian matrix in Normal distribution.
}
