% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normal.R
\name{normalMLE}
\alias{normalMLE}
\title{Get MLE estimate for Normal}
\usage{
normalMLE(x)
}
\arguments{
\item{x}{vector of data}
}
\value{
MLE estimates of mean and sd
}
\description{
Get MLE estimate for Normal
}
