% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normal.R
\name{normalPIT}
\alias{normalPIT}
\title{Calculate probability inverse transform for Normal distribution}
\usage{
normalPIT(x, theta)
}
\arguments{
\item{x}{vector of data}

\item{theta}{mle}
}
\value{
vector of probability inverse transformed values
}
\description{
Calculate probability inverse transform for Normal distribution
}
