% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normal.R
\name{normalScore}
\alias{normalScore}
\title{Score function for Normal dist}
\usage{
normalScore(x, theta)
}
\arguments{
\item{x}{vector of data}

\item{theta}{mle}
}
\value{
Score matrix
}
\description{
Score function for Normal dist
}
