% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normal.R
\name{observedHessianMatrixNormal}
\alias{observedHessianMatrixNormal}
\title{Calculate observed Hessian matrix for Normal distribution}
\usage{
observedHessianMatrixNormal(par)
}
\arguments{
\item{par}{a vector of mle parameters}
}
\value{
Observed Hessian matrix for Normal distribution
}
\description{
Calculate observed Hessian matrix for Normal distribution
}
