% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGLMGamma.R
\name{testGLMGamma}
\alias{testGLMGamma}
\title{Apply Goodness of Fit Test to the Residuals of a Generalized Linear Model}
\usage{
testGLMGamma(
  x,
  y,
  l,
  fit = NULL,
  start.value = NULL,
  control = NULL,
  method = "cvm"
)
}
\arguments{
\item{x}{is either a vector of length n or a design matrix with n rows and p columns.}

\item{y}{is a vector with the same number of observations or number of rows as x.}

\item{l}{a character vector indicating the link function that should be used for Gamma family. Some common
link functions for Gamma family are 'log' and 'inverse'. For more details see \code{\link{make.link}} from stats
package in R.}

\item{fit}{If you have already applied your generalized linear model using a link function for Gamma family,
instead of passing your data to the function, you can simply pass your fit object generated by
\code{\link{glm}} function. The default value of fit is NULL. If a fit object is provided, x, y, and l will be
ignored. The testGLMGamma function then will test if the residuals follow a Gamma distribution.
The software will check to make sure the fit object is returned from glm function and link function is indeed
from Gamma family. Also make sure to ask glm function to return the design matrix.
You can control this by setting x = TRUE in glm function.
To read more about this see the help documentation for \code{\link{glm}} function.}

\item{start.value}{TBD}

\item{control}{a list of parameters to control the fitting process in glm2 function. For more details, see \code{\link{glm.control}}.}

\item{method}{a character string to indicate which statistics to calculate.
The possible values are 'cvm' for Cramer-von Mises and 'ad' for Anderson-Darling.
The default value is 'cvm'.}
}
\value{
A list of two.
\itemize{
\item Statistic: The Cramer-von-Mises statistic.
\item pvalue: The approximated pvalue for the GoF test based on EDF.
}
}
\description{
This function can run the goodness of fit test to check if the residuals obtained from a generalized
linear model follows a Gamma distribution. Common link functions in glm function can be used here.
The function can take a fit object returned by glm functions or you can simply pass design matrix, x, and
response vector, y. The function returns Cramer-von Mises statistic along with approximate pvalue for the test.
}
\examples{
set.seed(123)
n <- 50
p <- 2
X <- matrix( rnorm(n*p, mean = 10, sd = 0.1), nrow = n, ncol = p)
b <- runif(p)
e <- rgamma(n, shape = 3)
y <- exp(X \%*\% b) * e
testGLMGamma(x=X,y,l = 'log')
myfit <- glm(y~X, family = Gamma('log'), x = TRUE)
testGLMGamma(fit = myfit)
}
