% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNormal.R
\name{testNormal}
\alias{testNormal}
\title{Apply Goodness of Fit Test for Normal Distribution}
\usage{
testNormal(
  x,
  ngrid = length(x),
  gridpit = FALSE,
  hessian = FALSE,
  method = "cvm"
)
}
\arguments{
\item{x}{A numeric vector of data values with length n.}

\item{ngrid}{The number of equally spaced points to discritize the (0,1) interval to estimate the covariance of the process.}

\item{gridpit}{A Boolean indicator. If TRUE, ngrid is ignored and (0,1) interval is divided based on probability inverse transformed
values. If FALSE (default value), (0,1) is divided into ngrid equally spaced points to estimate process.}

\item{hessian}{A Boolean indicator to control if Hessian matrix should be used in estimation of Fisher information matrix.
If False (the default value), the Fisher information matrix is estimated by the observed score function.
If TRUE, the Fisher information matrix is estimated by the observed Hessian Matrix.}

\item{method}{a character string to indicate which statistics to calculate.
The possible values are 'cvm' for Cramer-von Mises and 'ad' for Anderson-Darling.
The default value is 'cvm'.}
}
\value{
A list of two.
\itemize{
\item Statistic: The Cramer-von-Mises statistic.
\item pvalue: The approximated pvalue for the GoF test based on EDF.
}
}
\description{
Apply Goodness of Fit Test for Normal Distribution
}
\examples{
set.seed(123)
sim_data <- rnorm(n = 100)
testNormal(x = sim_data)
sim_data <- rgamma(100, shape = 3)
testNormal(x = sim_data)
}
