% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_exp.R
\name{expMLE}
\alias{expMLE}
\title{Calculate MLE of rate in Exponential dist.}
\usage{
expMLE(x)
}
\arguments{
\item{x}{a numeric vector of length n}
}
\value{
Numeric
}
\description{
Calculate MLE of rate in Exponential dist.
}
