% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_exp.R
\name{expScore}
\alias{expScore}
\title{Calculate Score Function for Exponential Distribution}
\usage{
expScore(x, theta)
}
\arguments{
\item{x}{a vector data}

\item{theta}{MLE estimate of data}
}
\value{
A matrix of nx1
}
\description{
Calculate Score Function for Exponential Distribution
}
