% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normal.R
\name{normalMLE}
\alias{normalMLE}
\title{Compute MLE estimate for Normal}
\usage{
normalMLE(x)
}
\arguments{
\item{x}{a numeric vector of length n}
}
\value{
a numeric vector of length two, containing MLE of parameters in Normal dist.
}
\description{
Compute MLE estimate for Normal
}
