% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normal.R
\name{normalPIT}
\alias{normalPIT}
\title{Compute probability inverse transform values for Normal distribution}
\usage{
normalPIT(x, theta)
}
\arguments{
\item{x}{a numeric vector of length n}

\item{theta}{a numeric vector of length two, containing MLE of parameters in Normal dist.}
}
\value{
a vector of length n containing the probability inverse transformed (PIT) values
}
\description{
Compute probability inverse transform values for Normal distribution
}
