% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normal.R
\name{normalScore}
\alias{normalScore}
\title{Compute score function for Normal dist}
\usage{
normalScore(x, theta)
}
\arguments{
\item{x}{a numeric vector of length n}

\item{theta}{a numeric vector of length two, containing MLE of parameters in Normal dist.}
}
\value{
Score matrix with n rows and two columns.
}
\description{
Compute score function for Normal dist
}
