% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testExponential.R
\name{testExponential}
\alias{testExponential}
\title{Apply Goodness of Fit Test for Exponential Distribution}
\usage{
testExponential(
  x,
  ngrid = length(x),
  gridpit = FALSE,
  hessian = FALSE,
  method = "cvm"
)
}
\arguments{
\item{x}{A numeric vector of data values with length n.}

\item{ngrid}{The number of equally spaced points to discritize the (0,1) interval to estimate the covariance of the process.}

\item{gridpit}{A Boolean indicator. If TRUE, ngrid is ignored and (0,1) interval is divided based on probability inverse transformed
values. If FALSE (default value), (0,1) is divided into ngrid equally spaced points to estimate process.}

\item{hessian}{A Boolean indicator. If TRUE, the Fisher information matrix is estimated by observed Hessian Matrix.
If False (the default value), the Fisher information matrix is estimated by observed score function.}

\item{method}{a character string to indicate which statistics to calculate.
The possible values are 'cvm' for Cramer-von Mises and 'ad' for Anderson-Darling.
The default value is 'cvm'.}
}
\value{
A list of two.
\itemize{
\item Statistic: The Cramer-von-Mises statistic.
\item pvalue: The approximated pvalue for the GoF test based on EDF.
}
}
\description{
Apply Goodness of Fit Test for Exponential Distribution
}
\examples{
set.seed(123)
sim_data <- rexp(n = 100)
testExponential(x = sim_data)
}
