% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testLMNormal.R
\name{testLMNormal}
\alias{testLMNormal}
\title{Apply Goodness of Fit Test to Residuals of a Linear Model}
\usage{
testLMNormal(
  x,
  y,
  fit = NULL,
  ngrid = length(y),
  gridpit = FALSE,
  hessian = FALSE,
  method = "cvm"
)
}
\arguments{
\item{x}{is either a vector of length n or a design matrix with n rows and p columns.}

\item{y}{is a vector with the same number of observations or number of rows as x.}

\item{fit}{If you have already applied your linear model, instead of passing your
data to the function, you can simply pass your fit object generated by
\code{\link{lm}} function. The default value of fit is NULL. If a fit object is provided, x and y will be
ignored. The testLMNormal function then will test if the residuals follow a Normal distribution.
The software will check to make sure the fit object is returned from lm function
Also make sure to ask lm function to return the design matrix and response variable used.
You can control this by setting x = TRUE and y = TRUE in lm function.

To read more about this see the help documentation for \code{\link{lm}} function.}

\item{ngrid}{The number of equally spaced points to discritize the (0,1) interval to estimate the covariance
of the process.}

\item{gridpit}{A boolean variable indicating if probability transformed values should be used to discritize
the (0,1) interval to estimate the covariance of the stochastic process.
If TRUE, ngrid is ignored and (0,1) interval is divided based on probability inverse transformed values.
If FALSE (default value), (0,1) is divided into ngrid equally spaced points to estimate process.}

\item{hessian}{A Boolean indicator to control if Hessian matrix should be used in estimation of Fisher information matrix.
If False (the default value), the Fisher information matrix is estimated by the observed score function.
If TRUE, the Fisher information matrix is estimated by the observed Hessian Matrix.}

\item{method}{a character string to indicate which statistics to calculate.
The possible values are 'cvm' for Cramer-von Mises and 'ad' for Anderson-Darling.
The default value is 'cvm'.}
}
\value{
A list of two.
\itemize{
\item Statistic: The Cramer-von-Mises statistic.
\item pvalue: The approximated pvalue for the GoF test based on EDF.
}
}
\description{
This function takes y as the response variable and x as the explanatory variable or design matrix
to test the normality of residuals in a linear model with the goodness of fit test.
It returns Cramer-von Mises statistic along with approximate pvalue for the test.
}
\examples{
set.seed(123)
n <- 50
p <- 5
xmatrix <- matrix( runif(n*p), nrow = n, ncol = p)
e <- runif(n)
b <- runif(p)
y <- xmatrix \%*\% b + e
testLMNormal(x = xmatrix, y)
# Or pass myfit object directly:
myfit <- lm(y ~ 0 + xmatrix, x = TRUE, y = TRUE)
testLMNormal(fit = myfit)
}
